Autotools build and one function per file layout
* Add files configure.ac, Makefile.am and src/Makefile.am with autotools
  configuration.

* Split udf_infusion.c into a number of separate files, one for each
  function.

* Put common definitions from udf_infusion.c to src/common.h.

* Add load.sql.sh and unload.sql.sh scripts for generating load.sql and
  unload.sql according to user-selected functions with:

    ./configure --enable-functions="..."

* Add LICENSE file with the original license text from udf_infusion.c.

* Add a number of additional files generated by autoreconf -i.
